+++ ipinfo/client.go
@@ -1,6 +1,7 @@
 package ipinfo
 
 import (
+	"context"
 	"encoding/json"
 	"fmt"
 	"io"
@@ -12,7 +13,7 @@ import (
 
 const (
 	defaultBaseURL   = "https://ipinfo.io/"
-	defaultUserAgent = "IPinfoClient/Go/2.1.1"
+	defaultUserAgent = "IPinfoClient/Go/2.2.0"
 )
 
 // A Client is the main handler to communicate with the IPinfo API.
@@ -63,12 +64,22 @@ func NewClient(
 	}
 }
 
-// NewRequest creates an API request. A relative URL can be provided in urlStr,
-// in which case it is resolved relative to the BaseURL of the Client. Relative
-// URLs should always be specified without a preceding slash.
-func (c *Client) NewRequest(urlStr string) (*http.Request, error) {
+// `newRequest` creates an API request. A relative URL can be provided in
+// urlStr, in which case it is resolved relative to the BaseURL of the Client.
+// Relative URLs should always be specified without a preceding slash.
+func (c *Client) newRequest(
+	ctx context.Context,
+	method string,
+	urlStr string,
+	body io.Reader,
+) (*http.Request, error) {
+	if ctx == nil {
+		ctx = context.Background()
+	}
+
 	u := new(url.URL)
 
+	// get final URL path.
 	if rel, err := url.Parse(urlStr); err == nil {
 		u = c.BaseURL.ResolveReference(rel)
 	} else if strings.ContainsRune(urlStr, ':') {
@@ -79,11 +90,13 @@ func (c *Client) NewRequest(urlStr string) (*http.Request, error) {
 		return nil, err
 	}
 
-	req, err := http.NewRequest("GET", u.String(), nil)
+	// get `http` package request object.
+	req, err := http.NewRequestWithContext(ctx, method, u.String(), body)
 	if err != nil {
 		return nil, err
 	}
 
+	// set common headers.
 	req.Header.Set("Accept", "application/json")
 	if c.UserAgent != "" {
 		req.Header.Set("User-Agent", c.UserAgent)
@@ -95,19 +108,22 @@ func (c *Client) NewRequest(urlStr string) (*http.Request, error) {
 	return req, nil
 }
 
-// Do sends an API request and returns the API response. The API response is
+// `do` sends an API request and returns the API response. The API response is
 // JSON decoded and stored in the value pointed to by v, or returned as an
 // error if an API error has occurred. If v implements the io.Writer interface,
 // the raw response body will be written to v, without attempting to first
 // decode it.
-func (c *Client) Do(req *http.Request, v interface{}) (*http.Response, error) {
+func (c *Client) do(
+	req *http.Request,
+	v interface{},
+) (*http.Response, error) {
 	resp, err := c.client.Do(req)
 	if err != nil {
 		return nil, err
 	}
 	defer resp.Body.Close()
 
-	err = CheckResponse(resp)
+	err = checkResponse(resp)
 	if err != nil {
 		// even though there was an error, we still return the response
 		// in case the caller wants to inspect it further
@@ -120,7 +136,8 @@ func (c *Client) Do(req *http.Request, v interface{}) (*http.Response, error) {
 		} else {
 			err = json.NewDecoder(resp.Body).Decode(v)
 			if err == io.EOF {
-				err = nil // ignore EOF errors caused by empty response body
+				// ignore EOF errors caused by empty response body
+				err = nil
 			}
 		}
 	}
@@ -130,8 +147,12 @@ func (c *Client) Do(req *http.Request, v interface{}) (*http.Response, error) {
 
 // An ErrorResponse reports an error caused by an API request.
 type ErrorResponse struct {
-	Response *http.Response // HTTP response that caused this error
-	Err      struct {
+	// HTTP response that caused this error
+	Response *http.Response
+
+	// Error structure returned by the IPinfo Core API.
+	Status string `json:"status"`
+	Err    struct {
 		Title   string `json:"title"`
 		Message string `json:"message"`
 	} `json:"error"`
@@ -143,10 +164,10 @@ func (r *ErrorResponse) Error() string {
 		r.Response.StatusCode, r.Err)
 }
 
-// CheckResponse checks the API response for errors, and returns them if
+// `checkResponse` checks the API response for errors, and returns them if
 // present. A response is considered an error if it has a status code outside
 // the 200 range.
-func CheckResponse(r *http.Response) error {
+func checkResponse(r *http.Response) error {
 	if c := r.StatusCode; 200 <= c && c <= 299 {
 		return nil
 	}